Header/banner test improvements
v0.2
* File renamed 'banner.spec.ts' to align with DOM object
* Test functions condensed to 3 due to performance issues
